feat: add CommandRunner and correction CommandLib
