--- practice2025/CommandLib/Class1.cs
@@ -1,4 +1,0 @@
-﻿public interface ICommand
-{
-    void Execute();
-}
