+++ practice2025/CommandLib/CommandLib.cs
@@ -1,0 +1,7 @@
+﻿namespace CommandLib
+{
+public interface ICommand
+{
+    void Execute();
+}
+}
